refactor: merged TheDataExperience.Default
